refs #1: Major improvement in standardization of projects, this project succeeds the following checks done by the "AnalyseProjects" app: Versioned, Folder structures, Filepaths, MainWin/form, AppIcon, AutoUpdating, AboutBox. Licensing not implemented and other checks must still be programmed to be done (in AnalyseProjects). Also started building TESTS (only two minor tests so far). Many other improvements not necessarily related to the standardization of projects.
git-svn-id: svn://localhost/sharedclasses/trunk@407 3f1363cb-6aae-0410-a37d-b56c5251702a

+++ TodoFileInterop.cs
@@ -88,7 +88,7 @@ namespace SharedClasses
 			{
 				ActionsInterop.DoAllActionsAndHandleError(
 					err => UserMessages.ShowErrorMessage(err),
-					delegate { File.Delete(FullFilePath); });
+					delegate { if (File.Exists(FullFilePath)) File.Delete(FullFilePath); });
 			}
 			if (createdTodoFiles.Contains(this))
 				createdTodoFiles.Remove(this);
@@ -99,7 +99,7 @@ namespace SharedClasses
 			string backupFilePath = GetDateFilenameNow();
 			bool saveSuccess = ActionsInterop.DoAllActionsAndHandleError(
 				(err) => UserMessages.ShowErrorMessage("Cannot save file: " + err),
-				delegate { File.WriteAllText(backupFilePath, File.ReadAllText(FullFilePath)); },
+				delegate { File.WriteAllText(backupFilePath, File.Exists(FullFilePath) ? File.ReadAllText(FullFilePath) : ""); },
 				//new FileInfo(FullFilePath).Compress(backupFilePath);
 				delegate { File.SetAttributes(backupFilePath, FileAttributes.System | FileAttributes.Hidden); },
 				delegate { File.WriteAllText(FullFilePath, _filecontent); });
@@ -168,6 +168,8 @@ namespace SharedClasses
 
 		public void Purge()
 		{
+			if (!File.Exists(FullFilePath))
+				return;
 			File.Move(FullFilePath, GetDateFilenameNow());
 		}
